refactor vault

+++ contracts/HedgepieVault.sol
@@ -14,17 +14,21 @@ contract HedgepieVault {
 
     address public immutable HedgepieToken;
     uint8 public constant blockEmission = 5;
-    uint public totalStake;
+    uint256 public totalStake;
     mapping(address => mapping(address => UserStake)) public userStake;
     struct UserStake {
-        uint start;
-        uint reward;
-        uint amount;
+        uint256 start;
+        uint256 reward;
+        uint256 amount;
     }
 
-    event Stake(address indexed _staker, address _token, uint _amount);
-    event Unstake(address indexed _unStaker, address _token, uint _amount);
-    event rewardClaim(address indexed _claimer, address _token, uint _amount);
+    event Stake(address indexed _staker, address _token, uint256 _amount);
+    event Unstake(address indexed _unStaker, address _token, uint256 _amount);
+    event rewardClaim(
+        address indexed _claimer,
+        address _token,
+        uint256 _amount
+    );
 
     constructor(address _hedgepie) {
         require(_hedgepie != address(0));
@@ -32,26 +36,38 @@ contract HedgepieVault {
         HedgepieToken = _hedgepie;
     }
 
-    function _getReward(address _token) private view returns(uint _reward) {
+    function _getReward(address _token) private view returns (uint256 _reward) {
         UserStake memory info = userStake[msg.sender][_token];
-        if(info.start > 0) {
-            uint blockDiff = block.number.sub(info.start).div(blockEmission);
-            _reward = totalStake > 0 ? FixedPoint.fraction( blockDiff.mul(info.amount), totalStake ).decode112with18() : 0;
+        if (info.start > 0) {
+            uint256 blockDiff = block.number.sub(info.start).div(blockEmission);
+            _reward = totalStake > 0
+                ? FixedPoint
+                    .fraction(blockDiff.mul(info.amount), totalStake)
+                    .decode112with18()
+                : 0;
         }
     }
 
-    function _checkUnstake(address _token, uint _amount) private view returns(bool) {
+    function _checkUnstake(address _token, uint256 _amount)
+        private
+        view
+        returns (bool)
+    {
         UserStake memory info = userStake[msg.sender][_token];
         return info.amount >= _amount;
     }
 
-    function _checkReward(address _token, uint _amount) private view returns(bool) {
+    function _checkReward(address _token, uint256 _amount)
+        private
+        view
+        returns (bool)
+    {
         UserStake memory info = userStake[msg.sender][_token];
         return info.reward.add(_getReward(_token)) >= _amount;
     }
 
-    function _stake(address _token, uint _amount) private {
-        IBEP20( _token ).safeTransferFrom(msg.sender, address(this), _amount);
+    function _stake(address _token, uint256 _amount) private {
+        IBEP20(_token).safeTransferFrom(msg.sender, address(this), _amount);
 
         UserStake memory info = userStake[msg.sender][_token];
         userStake[msg.sender][_token] = UserStake({
@@ -63,7 +79,7 @@ contract HedgepieVault {
         totalStake = totalStake.add(_amount);
     }
 
-    function _unstake(address _token, uint _amount) private {
+    function _unstake(address _token, uint256 _amount) private {
         UserStake memory info = userStake[msg.sender][_token];
         userStake[msg.sender][_token] = UserStake({
             amount: info.amount,
@@ -71,10 +87,10 @@ contract HedgepieVault {
             start: block.number
         });
 
-        IBEP20( _token ).safeTransfer(msg.sender, _amount);
+        IBEP20(_token).safeTransfer(msg.sender, _amount);
     }
 
-    function _claimReward(address _token, uint _amount) private {
+    function _claimReward(address _token, uint256 _amount) private {
         UserStake memory info = userStake[msg.sender][_token];
         userStake[msg.sender][_token] = UserStake({
             amount: info.amount.sub(_amount),
@@ -84,15 +100,19 @@ contract HedgepieVault {
 
         totalStake = totalStake.sub(_amount);
 
-        IBEP20( _token ).safeTransfer(msg.sender, _amount);
+        IBEP20(_token).safeTransfer(msg.sender, _amount);
     }
 
-    function getStake(address token) public view returns(UserStake memory info) {
+    function getStake(address token)
+        public
+        view
+        returns (UserStake memory info)
+    {
         info = userStake[msg.sender][token];
         info.reward = info.reward.add(_getReward(token));
     }
 
-    function stake(uint amount) public returns(bool) {
+    function stake(uint256 amount) public returns (bool) {
         require(amount > 0);
 
         _stake(HedgepieToken, amount);
@@ -101,9 +121,13 @@ contract HedgepieVault {
         return true;
     }
 
-    function stakeLP(address token, uint amount) public returns(bool) {
+    function stakeLP(address token, uint256 amount) public returns (bool) {
         require(amount > 0);
-        require(IPancakePair( token ).token0() == HedgepieToken || IPancakePair( token ).token1() == HedgepieToken, "Not LP token");
+        require(
+            IPancakePair(token).token0() == HedgepieToken ||
+                IPancakePair(token).token1() == HedgepieToken,
+            "Not LP token"
+        );
 
         _stake(token, amount);
 
@@ -111,17 +135,17 @@ contract HedgepieVault {
         return true;
     }
 
-    function unstake(address token, uint amount) public returns(bool) {
+    function unstake(address token, uint256 amount) public returns (bool) {
         require(amount > 0);
         require(_checkUnstake(token, amount), "Insufficient amount");
-        
+
         _unstake(token, amount);
 
         emit Unstake(msg.sender, token, amount);
         return true;
     }
 
-    function claimReward(address token, uint amount) public returns(bool) {
+    function claimReward(address token, uint256 amount) public returns (bool) {
         require(amount > 0);
         require(_checkReward(token, amount), "Insufficient amount");
